requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
onnx
onnxruntime
numpy
pillow
onnxscript